+++ src/lib/api/server.ts
@@ -2,8 +2,9 @@ import authRouter from './auth/auth.routes';
 import express from 'express';
 import userRouter from './users/users.routes';
 import onboardingRouter from './onboarding/onboarding.routes';
-import { createMap, gatherSummaryByDepartment } from '../utils/RMPScraper';
+//import { createMap, gatherSummaryByDepartment } from '../utils/RMPScraper';
 //import { gatherRMPSummary } from '../utils/RMPScraper';
+//import { department_list, department_professor_object } from '../utils/data/constants';
 
 const app = express();
 // ** needed to add express.json()
@@ -22,6 +23,7 @@ app.get('/test', (req, res) => {
 });
 
 app.listen('4001', async () => {
-  await gatherSummaryByDepartment(createMap(), 'american studies');
+  // TODO : call on the final function here
+  //await gatherSummaryByDepartment(inputMap, 'american studies'); --> this function gets called in the scraper itself
   console.log(`Listening to port 4001`);
 });
