+++ src/lib/utils/data/constants.ts
@@ -1,0 +1,377 @@
+interface department_professor_object_type {
+  /**
+   * @department is the placeholder for the key of the object, we use [] to specify that this is mutable
+   * @string[] is used to specify array of strings
+   */
+  [department: string]: string[];
+}
+
+const department_professor_object: department_professor_object_type = {
+  Administration: [
+    'Shant Shahrigian',
+    'Tony Liss',
+    'James(jim) Davis',
+    'Kara Pernicano',
+    'Sameeullah Sher',
+    'Paul Occhigrosso',
+    'Zeev Dagan',
+    'Warren Orange',
+    'Gregory Williams',
+    'Ashiwel Undieh',
+    'Tara Nachtigall',
+  ],
+  Advisor: [
+    'Valerie Poulolo',
+    'Arnaldo Melendez',
+    'Lavie Margolin',
+    'Dawn Marczak',
+    'Alice Shepard',
+    'Belinda Smith',
+    'Debra Kennedy',
+    'Nkem Nkem Stanley-Mbamlu',
+    'Marie Nazon',
+  ],
+  'African-American Studies': [
+    'Claude Brathwaite',
+    'Cheryl Sterling',
+    'Zaheer Ali',
+    'Heather Denyer',
+    'Herb Boyd',
+    'Timothy Mangin',
+    'Rebecca Wellington',
+    'Michelle Thompson',
+    'Gregory Baggett',
+    'Tanzeem Ajmiri',
+    'Paula Austin',
+    'Jeniffer Onyedum',
+    'Robert Gibbons',
+    'Farnia Noel',
+    'Micheal Gillepsie',
+    'Georgina Falu',
+    'Tyrene Wright',
+    'Abdul-Qasir Islam',
+  ],
+  'American Studies': [
+    'Kareen Williams',
+    'timothy Nicholson',
+    'Christopher Burrell',
+    'Judy Hilkey',
+    'Cody Nager',
+    'Hajoe Moderegger',
+  ],
+  Anthropology: [
+    'Bernando Chiachia M',
+    'Arthur Spears',
+    'Lindsay Parme',
+    'Cecillia Salvi',
+    'Mohammed Junaid',
+    'Simanique Moody',
+    'Melissa Maldonando-Salcedo',
+    'Kate Riley',
+    'Samuel Byrd',
+    'Kamal Soleimani',
+    'Stanley Thangaraj',
+    'Alice Baldwin Jones',
+    'Omnia Kalili',
+    'Monica Ortiz-Suloway',
+    'Jose Vazquez',
+    'Miriam Entin',
+    'Asha Samad-Matias',
+    'Carol Laderman',
+    'Melissa Zavala',
+    'William(bill) Askins',
+    'Andrea Ariza Garcia',
+    'Rafael Munia',
+    'Dane Ruffin',
+    'Lopez Joan Lopez',
+    'J Yarwood',
+    'Charles Rolph',
+    'D. Wall',
+    'Maggie Whitten',
+    'John Calagione',
+    'Stuart Wright',
+    'Shanelle Matthews',
+    'Marcela Restrepo',
+    'Robert Siebert',
+    'Adam Mikhail',
+    'Katherine Stefatos',
+    'Julie Lozano',
+    'Micheal Elster',
+    'Luis Nieto',
+    'Diane Sank',
+    'Stuart Rockefeller',
+    'Bruce Burnside',
+    'Matthew Reilly',
+    'Scott Schwartz',
+    'Theodor Maghrak',
+    'Charles Dolph',
+    'Branka Djuknic',
+    'Leah Mollin-Kling',
+    'Rocio Gil Martinez De Escobar',
+    'Miriam Fried',
+    'Mohammed Junaid',
+    'James Tolleson',
+  ],
+  Arabic: [
+    'Shareah Teleghani',
+    'Jeremy Randall',
+    'Rhoda Ismail',
+    'Karim Elhaies',
+    'Iman Issa',
+    'Mohammed Tkachmita',
+    'Mariam Elhaies',
+  ],
+  Architecture: [
+    'M.T. Chang',
+    'Jerrilyn Dodds',
+    'Marta Gutman',
+    'Lee Weintraub',
+    'Ali Hocek',
+    'W.Garrison McNeil',
+    'Alberto Foyo',
+    'Jeremy Edmiston',
+    'Peter Gisolfi',
+    'Julia Salcedo',
+    'Jane Kim',
+    'John Yurchyk',
+    'Kenneth Petrocca',
+    'Howard Duffy',
+    'Alfonso Oliva',
+    'Joshua Jow',
+    'Dexter Ciprian',
+    'Abdullah Khawarzad',
+    'Burt Goncalves',
+    'Frank Melendez',
+    'Cesare Birignani',
+    'Phillip Lee',
+    'Oleksandra Topolnytska',
+    'James Khamsi',
+    'Micheal King',
+    'Pablo Iglesias',
+    'Gonzalo Lopez',
+    'Kyle Dugdale',
+    'Damon Bolhassani',
+    'Brad Horn',
+    'Gordon Gebert',
+    'Sean Weiss',
+    'Athanasios Haritos',
+    'Robert Twombly',
+    'Eliana Dotan',
+    'Christian Volkmann',
+    'Timothy Collins',
+    'Johanna Dickson',
+    'Jessica Larson',
+    'Suzanne Strum',
+    'Olivia Vien',
+    'Maria Bermudez',
+    'Bradley Horn',
+    'Fabian Llonch',
+    'Nandini Bagchee',
+    'Antonio Di Oronzo',
+    'Ivan Rosa',
+    'Dominick Pilla',
+    'Lance Brown',
+    'Cassim Shephard',
+    'Irma Ostroff',
+    'Ahy Aydogan',
+    'June Williamson',
+    'Jason Kim',
+    'Demir Purisic',
+    'Paul Ruppert',
+    'John Cunningham',
+    'Tyler Survant',
+    'David Hotson',
+    'Jeffrey Roberson',
+    'Kaitlin Faherty',
+    'Jeremy Edmiston',
+    'Christian Martos',
+    'Brandt Graves',
+    'Jerome Haferd',
+    'Fran Leadon',
+    'Maria Berman',
+    'David Judelson',
+    'Alan Feigenberg',
+    'Denis Hoffman-Brandtt',
+    'Antonio Furgiuelle',
+    'Andrew Zago',
+    'Eduardo Alfonso',
+    'Hillary Brown',
+    'John Murray',
+    'Erik Forman',
+    'VIREN BRAHMBHATT',
+    'Beth Carliner',
+    'Mohammad Bolhassani',
+    'Alan Cation',
+    'Ruo Jia',
+  ],
+  //Art: [''],
+  //'Art History': [''],
+  //'Asian Studies': [''],
+  // Astronomy: [''],
+  // Biology: [''],
+  // Biomedical: [''],
+  // 'Biomedical Engineering': [''],
+  // Business: [''],
+  // 'Chemical Engineering': [''],
+  // Chemistry: [''],
+  // Chinese: [''],
+  // 'Civil Engineering': [''],
+  // Classics: [''],
+  // Communication: [''],
+  // 'Computer Science': [''],
+  // Design: [''],
+  // 'Earth Science': [''],
+  // Economics: [''],
+  // Education: [''],
+  // 'Electrical Engineering': [''],
+  // Engineering: [''],
+  // English: [''],
+  // 'English As A Second Language': [''],
+  // Environment: [''],
+  // 'Ethnic Studies': [''],
+  // Film: [''],
+  // Finance: [''],
+  // 'Fine Arts': [''],
+  // 'Foreign Languages & Literature': [''],
+  // French: [''],
+  // 'Freshman Year Experience': [''],
+  // Geography: [''],
+  // Geology: [''],
+  // German: [''],
+  // 'Graphic Arts': [''],
+  // 'Health Science': [''],
+  // Hindustani: [''],
+  // History: [''],
+  // Honors: [''],
+  // Hospitality: [''],
+  // Humanities: [''],
+  // 'Internation Service': [''],
+  // 'International Studies': [''],
+  // Italian: [''],
+  // Japanese: [''],
+  // 'Judaic Studies': [''],
+  // Languages: [''],
+  // Latin: [''],
+  // 'Latin American Studies': [''],
+  // Law: [''],
+  // 'Liberal Studies': [''],
+  // Libraries: [''],
+  // 'Library & Information Science': [''],
+  // Literature: [''],
+  // Management: [''],
+  // Marketing: [''],
+  // Mathematics: [''],
+  // 'Mechanical Engineering': [''],
+  // 'Media Arts': [''],
+  // Medicine: [''],
+  // Music: [''],
+  // Nursing: [''],
+  // Philosophy: [''],
+  // Physics: [''],
+  // 'Political Science': [''],
+  // Portugese: [''],
+  // Psychology: [''],
+  // 'Public Policy': [''],
+  // Science: [''],
+  // 'Social Science': [''],
+  // Sociology: [''],
+  // Spanish: [''],
+  // Speech: [''],
+  // Theater: [''],
+  // 'Urban Design & Development': [''],
+  // 'Urban Landscape': [''],
+  // "Women's Studies": [''],
+  // 'World Civilizations': [''],
+  // 'World Humanities': [],
+};
+
+const department_list: string[] = [
+  'Administration',
+  'Advisor',
+  'African-American Studies',
+  'American Studies',
+  'Anthropology',
+  'Arabic',
+  'Architecture',
+  // 'Art',
+  // 'Art History',
+  // 'Asian Studies',
+  // 'Astronomy',
+  // 'Biology',
+  // 'Biomedical',
+  // 'Biomedical Engineering',
+  // 'Business',
+  // 'Chemical Engineering',
+  // 'Chemistry',
+  // 'Chinese',
+  // 'Civil Engineering',
+  // 'Classics',
+  // 'Communication',
+  // 'Computer Science',
+  // 'Design',
+  // 'Earth Science',
+  // 'Economics',
+  // 'Education',
+  // 'Electrical Engineering',
+  // 'Engineering',
+  // 'English',
+  // 'English As A Second Language',
+  // 'Environment',
+  // 'Ethnic Studies',
+  // 'Film',
+  // 'Finance',
+  // 'Fine Arts',
+  // 'Foreign Languages & Literature',
+  // 'French',
+  // 'Freshman Year Experience',
+  // 'Geography',
+  // 'Geology',
+  // 'German',
+  // 'Graphic Arts',
+  // 'Health Science',
+  // 'Hindustani',
+  // 'History',
+  // 'Honors',
+  // 'Hospitality',
+  // 'Humanities',
+  // 'Internation Service',
+  // 'International Studies',
+  // 'Italian',
+  // 'Japanese',
+  // 'Judaic Studies',
+  // 'Languages',
+  // 'Latin',
+  // 'Latin American Studies',
+  // 'Law',
+  // 'Liberal Studies',
+  // 'Libraries',
+  // 'Library & Information Science',
+  // 'Literature',
+  // 'Management',
+  // 'Marketing',
+  // 'Mathematics',
+  // 'Mechanical Engineering',
+  // 'Media Arts',
+  // 'Medicine',
+  // 'Music',
+  // 'Nursing',
+  // 'Philosophy',
+  // 'Physics',
+  // 'Political Science',
+  // 'Portugese',
+  // 'Psychology',
+  // 'Public Policy',
+  // 'Science',
+  // 'Social Science',
+  // 'Sociology',
+  // 'Spanish',
+  // 'Speech',
+  // 'Theater',
+  // 'Urban Design & Development',
+  // 'Urban Landscape',
+  // "Women's Studies",
+  // 'World Civilizations',
+  // 'World Humanities',
+];
+
+export { department_professor_object, department_professor_object_type, department_list };
